# Add reactive programming and major optimizations