--- a/src/pyventus/__init__.py
+++ b/src/pyventus/__init__.py
@@ -2,8 +2,9 @@
 
 __version__ = "0.5.0"
 
-from .core.exceptions import PyventusException
+from .core.exceptions import PyventusException, PyventusImportException
 
 __all__ = [
     "PyventusException",
+    "PyventusImportException",
 ]