--- a/src/pyventus/core/exceptions/__init__.py
+++ b/src/pyventus/core/exceptions/__init__.py
@@ -1,1 +1,2 @@
 from .pyventus_exception import PyventusException
+from .pyventus_import_exception import PyventusImportException