--- a/src/pyventus/core/loggers/logger.py
+++ b/src/pyventus/core/loggers/logger.py
@@ -1,4 +1,4 @@
-from ..utils import attributes_repr, formatted_repr
+from ..utils import attributes_repr, formatted_repr, summarized_repr
 from .stdout_logger import StdOutLogger
 
 
@@ -6,16 +6,30 @@ class Logger:
     """A custom logger class that wraps the `StdOutLogger` and provides additional functionality."""
 
     # Attributes for the Logger
-    __slots__ = ("__name", "__debug")
+    __slots__ = ("__source", "__debug")
 
-    def __init__(self, name: str | None = None, debug: bool = False):
+    def __init__(self, source: str | type | object | None = None, debug: bool = False):
         """
         Initialize an instance of `Logger`.
 
-        :param name: The name of the logger instance.
+        :param source: The source of the log message, which can be a string, type, or object. Defaults to None.
         :param debug: A flag indicating whether debug mode is enabled.
         """
-        self.__name: str | None = name
+        # Variable to hold the name of the source.
+        source_name: str | None
+
+        # Determine the name of the source based on its type.
+        if source is None:
+            source_name = None
+        elif isinstance(source, str):
+            source_name = source
+        elif isinstance(source, type):
+            source_name = source.__name__
+        else:
+            source_name = summarized_repr(source)
+
+        # Store the determined source name and debug flag.
+        self.__source: str | None = source_name
         self.__debug = debug
 
     def __repr__(self) -> str:  # pragma: no cover
@@ -27,7 +41,7 @@ class Logger:
         return formatted_repr(
             instance=self,
             info=attributes_repr(
-                name=self.__name,
+                source=self.__source,
                 debug=self.__debug,
             ),
         )
@@ -49,7 +63,7 @@ class Logger:
         :param action: The action or method associated with the log. Defaults to None.
         :return: None.
         """
-        StdOutLogger.error(msg=msg, name=self.__name, action=action)
+        StdOutLogger.error(msg=msg, source=self.__source, action=action)
 
     def warning(self, msg: str, action: str | None = None) -> None:
         """
@@ -59,7 +73,7 @@ class Logger:
         :param action: The action or method associated with the log. Defaults to None.
         :return: None.
         """
-        StdOutLogger.warning(msg=msg, name=self.__name, action=action)
+        StdOutLogger.warning(msg=msg, source=self.__source, action=action)
 
     def info(self, msg: str, action: str | None = None) -> None:
         """
@@ -69,7 +83,7 @@ class Logger:
         :param action: The action or method associated with the log. Defaults to None.
         :return: None.
         """
-        StdOutLogger.info(msg=msg, name=self.__name, action=action)
+        StdOutLogger.info(msg=msg, source=self.__source, action=action)
 
     def debug(self, msg: str, action: str | None = None) -> None:
         """
@@ -79,4 +93,4 @@ class Logger:
         :param action: The action or method associated with the log. Defaults to None.
         :return: None.
         """
-        StdOutLogger.debug(msg=msg, name=self.__name, action=action)
+        StdOutLogger.debug(msg=msg, source=self.__source, action=action)