--- a/src/pyventus/core/processing/__init__.py
+++ b/src/pyventus/core/processing/__init__.py
@@ -1,0 +1,1 @@
+from .processing_service import ProcessingService, ProcessingServiceCallbackType