--- a/src/pyventus/core/processing/asyncio/__init__.py
+++ b/src/pyventus/core/processing/asyncio/__init__.py
@@ -1,0 +1,1 @@
+from .asyncio_processing_service import AsyncIOProcessingService