--- a/src/pyventus/core/processing/celery/__init__.py
+++ b/src/pyventus/core/processing/celery/__init__.py
@@ -1,2 +1,1 @@
-from .celery_payload import CeleryPayload
 from .celery_processing_service import CeleryProcessingService