--- a/src/pyventus/core/processing/celery/celery_payload.py
+++ b/src/pyventus/core/processing/celery/celery_payload.py
@@ -1,4 +1,4 @@
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 from typing import Any
 
 from ..processing_service import ProcessingServiceCallbackType
@@ -16,11 +16,3 @@ class CeleryPayload:
 
     kwargs: dict[str, Any]
     """Keyword arguments to be passed to the callback."""
-
-    def to_json(self) -> dict[str, Any]:
-        """
-        Convert the payload to a JSON-compatible dictionary.
-
-        :return: A JSON-compatible dictionary representing the payload.
-        """
-        return asdict(self)