--- a/src/pyventus/core/processing/executor/__init__.py
+++ b/src/pyventus/core/processing/executor/__init__.py
@@ -1,0 +1,1 @@
+from .executor_processing_service import ExecutorProcessingService