--- a/src/pyventus/events/emitters/celery/__init__.py
+++ b/src/pyventus/events/emitters/celery/__init__.py
@@ -1,1 +1,0 @@
-from .celery_event_emitter import CeleryEventEmitter