## Initial Checks

- [X] The code compiles successfully without any errors or warnings.
- [X] I have personally tested these changes on my local environment.
- [X] I have ensured that relevant documentation has been added or updated.
- [X] I adhere to the project structure and code standards defined in the documentation.
- [X] My code follows the established coding style guidelines.
- [X] I have added tests for the new code (if applicable).
- [X] All existing tests have passed successfully.

## Description

&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;This pull request introduces significant improvements to Pyventus, including the addition of the **reactive programming** paradigm and **major optimizations** to the previous event-driven features.

#### Breaking Changes

- The `typing-extensions` package is now a required dependency for Pyventus. This dependency has been added to support advanced typing features in older versions of Python.

- All previous event-driven features must now be imported from the inner package `pyventus.events` instead of directly from `pyventus`. These changes were necessary due to the integration of the reactive programming paradigm. This series of refactors were implemented not only to ensure an organized codebase but also to establish a clear boundary between event handling and reactive programming features, enabling optimized imports based on the required paradigm.

- The inheritance structure of the `EventEmitter` has been replaced with composition using the new `ProcessingService` interface. This interface establishes a common ground for processing calls, such as event emissions, in a decoupled and flexible manner. The `EventEmitter` is now a concrete class that requires a `ProcessingService` instance, referred to as `event_processor`, for initialization. These changes not only preserve the overall behavior and workflow of the `EventEmitter` but also enhance its modularity and flexibility.

- The `EventLinker` class has experienced multiple method renames and return type modifications to align with the new redesigned codebase, enhancing the API for improved usability and intuitiveness. Below is a detailed list of the breaking changes made to the `EventLinker` class:
	- The inner class `EventLinkageWrapper` was renamed to `EventLinkerSubCtx` for consistency with the new concept of a subscription context and now extends from the base class `SubscriptionContext`. The overall workflow remains roughly the same but was reworked to align with its base class and introduce optimizations based on user needs.
	- The method `get_events()` now returns a `set` of all registered events instead of a `list` with non-duplicated events.
	- The method `get_event_handlers()` was renamed to `get_subscribers()` for consistency, and it now returns a `set`of all registered subscribers instead of a `list` with non-duplicated subscribers.
	- The method `get_events_by_event_handler()` was renamed to `get_events_from_subscribers()` for consistency. It now returns a `set` of events associated with the specified subscribers instead of a `list` of non-duplicated events associated with the specified subscribers. This method now also supports retrieving events from multiple subscribers instead of only one at a time.
	- The method `get_event_handlers_by_events()` was renamed to `get_subscribers_from_events()` for consistency. It now returns a `set` of subscribers associated with the specified events instead of a `list` of non-duplicated subscribers associated with the provided events. Additionally, a new flag called `pop_onetime_subscribers` was added to remove and return those subscribers that are one-time subscriptions.
	- The method `unsubscribe()` was renamed to `remove()` for consistency with the new concept of subscription and the encapsulation of the unsubscription process through the `teardown_callback`. This method now allows you to remove one event and subscriber from the registry at a time instead of multiple events of the given subscriber.
	- The method `remove_event_handler()` was renamed to `remove_subscriber()` for consistency.
	- Parameters named `event_handler` were renamed to `subscriber` for consistency.

- The `EventHandler` has been refactored from a class to an interface, outlining the overall workflow and essential protocols for event handling. However, the previous implementation of the `EventHandler` has been transitioned to one of its concrete classes named `EventSubscriber`. This new `EventSubscriber` class not only implements the `EventHandler` interface but also combines it with the `Subscription` base class, providing a convenient way to both handle event responses and manage the subscription lifecycle.

#### Added

- Added a reactive programming module to expand Python's event-driven capabilities for data-oriented processes and the ability to react efficiently to state changes over time.
	- Added the `Observable` base class, which defines a lazy push-style notification mechanism for streaming data to subscribers.
	- Added the `ObservableTask` class, an observable subclass that encapsulates a unit of work and offers a mechanism for streaming its results reactively.
	- Added the `as_observable_task()` decorator to easily convert any given callable into an observable task.
	- Added the `Observer` interface, which defines the overall workflow and essential protocols for responding to notifications from an observable.
	- Added the `Subscriber` class, which combines the `Observer` interface with the `Subscription` base class to provide a convenient way to respond to state changes emitted by an observable and manage the subscription lifecycle.

- Added the `Unsubscribable` interface, which provides a standardized method for objects to unsubscribe from a source and release any associated resources.

- Added the `Subscription` base class to simplify subscription management and resource cleanup with a `teardown_callback` that is called during unsubscription.

- Added the `SubscriptionContext` base class, which defines the overall workflow for subscription contexts, allowing the user to define step-by-step the object that will later be subscribed to the specified source.

- Added a new package global exception called `PyventusImportException`, which is a custom Pyventus exception for handling missing imports within the library.

- Added the `MultiBidict` data structure, a generic multikeyed, multivalued bidirectional dictionary that offers a flexible mapping structure for efficient lookups, updates, and deletions of keys and their corresponding values.

- Introduced the `ProcessingService` interface to define flexible execution strategies for various use cases. This release includes the following concrete implementations:
	- `AsyncIOProcessingService`: A processing service that utilizes the `AsyncIO` framework to handle the execution of calls.
	- `CeleryProcessingService`: A processing service that utilizes the `Celery` framework to handle the execution of calls.
	- `ExecutorProcessingService`: A processing service that utilizes the Python's `Executor` to handle the execution of calls.
	- `FastAPIProcessingService`: A processing service that uses FastAPI's `BackgroundTasks` to handle the execution of calls.
	- `RedisProcessingService`: A processing service that utilizes the `Redis Queue` framework to handle the execution of calls.

- Added the `CallableWrapper` class to encapsulate callables and provide a unified asynchronous interface for their execution.

- Added new features to the `EventLinker` class, including the following methods:
	- Introduced the `get_valid_subscriber()` method for a centralized mechanism to validate event subscribers.
	- Added the `is_empty()` method to efficiently check if the main registry is empty.
	- Added the `get_event_count()` method to return the total number of events in the registry.
	- Added the `get_subscriber_count()` method to return the total number of subscribers in the registry.
	- Added the `get_event_count_from_subscriber()` method to return the number of events for a specific subscriber.
	- Added the `get_subscriber_count_from_event()` method to return the number of subscribers for a specific event.
	- Added the `contains_event()` method to check if a specific event is present in the registry.
	- Added the `contains_subscriber()` method to check if a specific subscriber is present in the registry.
	- Added the `are_linked()` method to determine if a specific event is linked to a given subscriber.
	- Introduced the `stateful_subctx` parameter in the `once()` and `on()` methods to configure the `EventLinkerSubCtx` behavior and optimize the subscription context based on user needs.

- Introduced the `EventSubscriber` class, which combines event handling capabilities with subscription lifecycle management.

- A new `benchmarks` package has been added to the `tests` directory for performance evaluation of Pyventus. This release introduces the `EventEmitterBenchmark`, which measures the efficiency of the `EventEmitter` in handling event emissions.

- Added a set of utilities for creating preconfigured event emitter instances, making the setup process easier. These utilities also provide retro compatibility with the previous class-based design.

#### Changed

- Enhanced the `emit()` method in the `EventEmitter` to support the emission of global events (`...`).

- Moved callback utilities to a dedicated module within `pyventus.core` for improved organization and reusability.

- Standardized the structure of Python classes and their representation, including the use of the `@override()` decorator for consistency and `mypy` static type checking.

- Standardized the structure of the `pyproject.toml` file.

- Enhanced Pyventus logs by adding process and thread IDs for better debugging.

- Switched from the `Black` formatter to `Ruff` for improved development efficiency and enhanced code quality.

- Refactored all project docstrings to follow a standardized format, enhancing consistency and clarity in the documentation.

- Upgraded several development dependencies in the `pyproject.toml`, including `pytest-asyncio` from version `0.21.0` to `0.24.0`, to enable global configuration of the `asyncio_mode`.

- Simplified the `EventCallbackType` type alias by removing the unnecessary `ParamSpec`.

- Refactored the test suite to improve validation across all package features, ensuring correctness and achieving 100% code coverage.


#### Optimized

- The time complexity of the `emit()` method in the `EventEmitter` class has been significantly optimized. It has been reduced from $O(N \cdot T)$ to $O(N \cdot L_{avg})$, where:

	- $N$: Represents the total number of subscribers involved in the event emission, including those registered to the global event and those specific to the event being emitted.
	- $T$: Denotes the total number of elements in the event linker registry, encompassing all subscriber lists combined. In mathematical terms, $T$ is defined as $\sum_{e}^{E} S_{e}$, where $S_{e}$ corresponds to the number of subscribers for the event $e$, and $E$ encompasses all events.
 	- $L_{avg}$: Represents the average number of events linked to one-time subscribers during event emission.

      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;This optimization results from the more efficient management of one-time subscribers by the event linker during event emission. Instead of traversing the entire event linker registry ($T$) to remove each one-time subscriber involved in the event emission ($N$), it now iterates solely through the linked events ($L$). This improvement is particularly beneficial when all subscribers engaged in the event emission are one-time subscriptions. However, in scenarios without one-time subscribers during an event emission, both implementations have the same time complexity of $O(N)$.

- Major optimizations have been implemented for the `EventLinker` class through the integration of the `MultiBidict` data structure. This data structure is a multikeyed, multivalued bidirectional dictionary implementation that enables efficient lookups, updates, and deletions of events and their corresponding subscribers. Despite utilizing a bidirectional mapping structure, its memory footprint remains minimal due to the use of references between keys and values instead of duplication, which limits the impact to the additional dictionary and set data structures.

	- The time complexity of the method `get_subscribers()`, previously known as `get_event_handlers()`, has been significantly reduced from $O(T)$ to $O(S)$, where:
		- $T$: Denotes the total number of elements in the event linker registry, encompassing all subscriber lists combined.
		- $S$: Corresponds to the total number of registered subscribers without duplications.

	- The time complexity of the `get_events_from_subscribers()`, previously known as `get_events_by_event_handler()`, has been reduced from $O(T)$ to $O(L)$, where:
		- $T$: Denotes the total number of elements in the event linker registry, encompassing all subscriber lists combined.
		- $L$: Represents the total number of events linked to the given subscriber.

          &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;It is important to note that the new method of `get_events_from_subscribers()` can receive not only one subscriber as before but also multiple subscribers to retrieve the associated events. In cases where multiple subscribers are provided, the time complexity of the overall method would be $O(N \cdot L_{avg})$, where $N$ corresponds to the number of distinct subscribers for which to query the linked events, and $L_{avg}$ represents the average number of events associated with each subscriber.

	- The time complexity of the method `remove_subscriber()`, previously known as `remove_event_handler()`, has been significantly reduced from $O(T)$ to $O(L)$, where:
		- $T$: Denotes the total number of elements in the event linker registry, encompassing all subscriber lists combined.
		- $L$: Represents the total number of events linked to the given subscriber.

 	- The time complexity of the method `remove()`, previously known as `unsubscribe()`, has been enhanced from $O(S)$ to a constant time complexity of $O(1)$, where $S$ represents the size of the subscriber list registered for the given event. It is important to note that while the previous comparison was made based on the `unsubscribe()` method handling just one event, under similar conditions to the `remove()` method, the previous implementation actually supported removing the subscriber from multiple events. The achievement of the new $O(1)$ time complexity is based on the efficiency of `set` operations in Python.

- Introduced `__slots__` in several classes to optimize memory usage and enhance attribute access speed.

#### Benchmarks

&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Finally, to provide a quick visualization of the overall improvements and illustrate the time complexity enhancements of this upcoming release, a series of benchmarks were conducted. These benchmarks were specifically designed for the event emission process, as it encompasses all event-driven features and provides a clear overview of the improvements. Following this, a detailed explanation of the benchmarks and their results is presented.

- **Methodology:** The benchmarks utilized a volume testing approach to assess how different subscription counts affect the event emission time.

- **Environment:** The benchmarks were conducted in the following environment:

	- **Operating System**: Windows 11 64bit (v10.0.22631)
	- **CPU**: AMD Ryzen 5 2600, 3400.0 MHz, 6 cores, 12 logical processors
	- **Total Memory**: 16 GB
	- **Python Version**: 3.12.7

- **Setup:** The benchmark setup consisted of two key components: a main Python script that managed the overall workflow of the benchmarks and the `EventEmitterBenchmark` class, which was essential for standardizing and organizing the performance tests. Following this, the main workflow of the benchmarks will be outlined, including the required packages and the adjustable settings available for different tests.

    <details markdown="1" class="quote">
    <summary>Benchmark Workflow and Settings (<code>main.py</code>)</summary>
    
    Before running the benchmarks, make sure all necessary packages are installed. You can install them using the following `pip` command:
    
	```console
	pip install pyventus[tests]
	```

	Additionally, an extra package is required for visualizing the results: `matplotlib`. You can also install it with a `pip` command as follows:
    
	```console
	pip install matplotlib
	```
    
    Once the packages are installed, you can configure the benchmarks in the `main.py` script as needed and execute it. The script will first install Pyventus `v0.6.0`, run the benchmarks in a separate process, and then uninstall it. Next, it will install the current version of Pyventus and run the benchmarks again in another process. Finally, the results for each version will be saved in a `JSON` file, and plots will be generated based on the benchmark reports.
    
	```Python title="main.py" linenums="1"
	import gc
	import os
	import subprocess
	import sys
	from concurrent.futures import ProcessPoolExecutor
	from dataclasses import asdict
	from json import dumps

	import matplotlib.pyplot as plt
	import numpy as np

	def run_event_emitter_benchmarks():
		from tests.benchmarks.event_emitter_benchmark import EventEmitterBenchmark

		print(f"Pyventus v{EventEmitterBenchmark.PYVENTUS_VERSION}\n")
		print("Starting benchmarks...")

		# Define a list of benchmarks to run with different configurations.
		benchmarks: list[EventEmitterBenchmark] = [
			# Benchmark with SINGLE event subscription mode and NONE one-time subscriptions.
			EventEmitterBenchmark(
				event_subscription_mode=EventEmitterBenchmark.EventSubscriptionMode.SINGLE,
				onetime_subscription_mode=EventEmitterBenchmark.OneTimeSubscriptionMode.NONE,
				subscription_sizes=[100, 500, 1000, 5000, 10000],
				num_repeats=5,
				num_executions=1250,
			),			
			# Benchmark with SINGLE event subscription mode and ALL one-time subscriptions.
			EventEmitterBenchmark(
				event_subscription_mode=EventEmitterBenchmark.EventSubscriptionMode.SINGLE,
				onetime_subscription_mode=EventEmitterBenchmark.OneTimeSubscriptionMode.ALL,
				subscription_sizes=[100, 500, 1000, 5000, 10000],
				num_repeats=5,
				num_executions=1250,
			),
			# Benchmark with ALL event subscription mode and ALL one-time subscriptions.
			EventEmitterBenchmark(
				event_subscription_mode=EventEmitterBenchmark.EventSubscriptionMode.ALL,
				onetime_subscription_mode=EventEmitterBenchmark.OneTimeSubscriptionMode.ALL,
				subscription_sizes=[100, 500, 1000, 5000, 10000],
				num_repeats=5,
				num_executions=1250,
			),
			# Benchmark with RANDOM event subscription mode and RANDOM one-time subscriptions.
			EventEmitterBenchmark(
				event_subscription_mode=EventEmitterBenchmark.EventSubscriptionMode.RANDOM,
				onetime_subscription_mode=EventEmitterBenchmark.OneTimeSubscriptionMode.RANDOM,
				subscription_sizes=[100, 500, 1000, 5000, 10000],
				num_repeats=3,
				num_executions=500,
			),
		]

		# Initialize an empty list to store reports generated by the benchmarks.
		reports: list[EventEmitterBenchmark.Report] = []

		# Iterate through each benchmark and execute it.
		for benchmark in benchmarks:
			reports.append(benchmark())

		# Return the list of reports.
		return reports

	def plot_event_emitter_benchmark_comparison(
		title: str, subscription_sizes: list[int], benchmark_bars: dict[tuple[str, str], list[float]], output_file: str
	) -> None:
…